+++ packages/docusaurus/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @to-skills/docusaurus
 
+## 0.2.19
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @to-skills/core@1.1.0
+
 ## 0.2.18
 
 ### Patch Changes